requirements.txt
# Django and core dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0

# Static files
whitenoise>=6.5,<7.0

# Environment variables
python-dotenv>=1.0,<2.0

# WSGI server for production
gunicorn>=21.0,<22.0

# For MySQL (if used locally)
# mysqlclient>=2.2,<3.0

# For report generation
reportlab>=4.0,<5.0

# Additional utilities
Pillow>=10.0,<11.0
